provide last updated time

+++ src/main.rs
@@ -87,7 +87,7 @@ pub struct User {
     pub message_count: Option<u64>,
     pub rank: i64,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub last_updated: Option<u128>,
+    pub last_updated: Option<i64>,
 }
 
 #[derive(serde::Deserialize, serde::Serialize)]
